+++ enterprise/toggles.py
@@ -1,0 +1,36 @@
+"""
+Waffle toggles for enterprise features within the LMS.
+"""
+
+from edx_toggles.toggles import WaffleFlag
+
+ENTERPRISE_NAMESPACE = 'enterprise'
+ENTERPRISE_LOG_PREFIX = 'Enterprise: '
+
+# .. toggle_name: enterprise.TOP_DOWN_ASSIGNMENT_REAL_TIME_LCM
+# .. toggle_implementation: WaffleFlag
+# .. toggle_default: False
+# .. toggle_description: Enables top-down assignment
+# .. toggle_use_cases: open_edx
+# .. toggle_creation_date: 2023-09-15
+TOP_DOWN_ASSIGNMENT_REAL_TIME_LCM = WaffleFlag(
+    f'{ENTERPRISE_NAMESPACE}.top_down_assignment_real_time_lcm',
+    __name__,
+    ENTERPRISE_LOG_PREFIX,
+)
+
+
+def top_down_assignment_real_time_lcm():
+    """
+    Returns whether top-down assignment and real time LCM feature flag is enabled.
+    """
+    return TOP_DOWN_ASSIGNMENT_REAL_TIME_LCM.is_enabled()
+
+
+def enterprise_features():
+    """
+    Returns a dict of enterprise Waffle-based feature flags.
+    """
+    return {
+        'top_down_assignment_real_time_lcm': top_down_assignment_real_time_lcm(),
+    }